requirements.txt
fastapi==0.116.1
pydantic==2.11.7
python-dotenv==1.1.1
requests==2.32.5
uvicorn==0.35.0
psycopg2==2.9.10
docx==0.2.4
lxml==6.0.1
pillow==11.3.0
pdfplumber==0.9.0